+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 ## azure-maps-geolocation-control Changelog
 
+
+<a name="0.0.2"></a>
+# 0.0.2 (2022-09-26)
+
+**Features**
+
+- Modified control to try high accuracy first, then fallback to low accuracy.
+- Bug fix: Updating of map position when tracking (was only working once).
+- Big fix: Adjusted marker so it stays flat on the map when map is pitched.
+
 <a name="0.0.1"></a>
 # 0.0.1 (2020-08-21)